covidTracing/location/serializers.py
from rest_framework import serializers

class LocationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Location
        fields = (
            'id',
            'location',
            'address',
            'city',
        )